feat(OpenCV): Starter Kit + C# Example
The starter kit includes all OpenCV files we need.

The examples folder includes the two provided by the german instructor for Python and C++.
It also includes my attempt at re-creating the C++ example, but that's a bust.

The C# example replicates the effects of the examples given to us except for mouse clicks.