feat(VideoFeed): Initial Setup
I added a function called SimpleVideoFeed()

Calling that instead of DetectCups() in main will open up a window with the video feed from the laptop camera.

This will be our base and reference for implementing video for the cups.

I will be using this simple video feed for testing mouse clicks.