refactor(NewMain): CV_Main & CV_Program
CV_Program will now hold most of the code, similar to how Program.cs used to.

Main is in CV_Main which calls the appropriate function from CV_Program.

For backup/reference, the old Program.cs is still around.

We can make a CV_Cup class as soon as it's needed.