fix(BoxSplit): Assymetric Split (For 2 Only)
When we split a big box into two, we take the position and width of the box from the previous frame into account to split it accordingly so that one part can be wide and the other small.

Run this with Sample 4 to see it happen when cup number two (which is infront of 3) is moved to the right of the screen.